+++ src/commands/core/botinfo.js
@@ -79,9 +79,7 @@ module.exports = class CMD extends Command {
             )
             .addField(
                 "🔗 Useful links:",
-                `> [Support server](${await message.client.config.supportGuildInviteReal(
-                    this.client
-                )})\n` +
+                `> [Support server](${message.client.config.supportGuildInvite})\n` +
                     "> [GitHub](https://github.com/Welcome-Bot/welcome-bot/)\n" +
                     "> [Privacy policy](https://welcome-bot.github.io/docs/privacy-policy.html) and [Terms of service](https://welcome-bot.github.io/docs/terms.html)\n" +
                     "> [Documentation](https://welcome-bot.github.io/docs)",