+++ src/commands/core/ping.js
@@ -3,6 +3,7 @@
  * Copyright (c) 2021 The Welcome-Bot Team and Contributors
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
+//eslint-disable-next-line no-unused-vars
 const { Embed, Command } = require("../../classes");
 module.exports = class CMD extends Command {
     constructor(client) {
@@ -15,11 +16,13 @@ module.exports = class CMD extends Command {
                 disabled: false,
                 cooldown: 5,
                 category: "Core",
+                supportsSlash: true,
             },
             client
         );
     }
 
+    //eslint-disable-next-line no-unused-vars
     execute({ message }, t) {
         const msg = `${t("misc:pong")} ${message.author}\n${t(
             "misc:webheart"
@@ -34,12 +37,14 @@ module.exports = class CMD extends Command {
         });
     }
 
+    //eslint-disable-next-line no-unused-vars
     async run({ interaction }, t) {
         const msg = `${t("misc:pong")} ${interaction.member.user}\n${t(
             "misc:webheart"
         )}: ${interaction.client.ws.ping}ms.\n`;
-        await interaction.reply(msg + `Getting roundtrip latency`);
-        const sent = await interaction.fetchReply();
+        const sent = await interaction.followUp(
+            msg + `Getting roundtrip latency`
+        );
         sent.edit(
             msg +
                 `${t("misc:latency")}: ${