+++ src/commands/economy/use.js
@@ -10,7 +10,7 @@ module.exports = class CMD extends Command {
         super(
             {
                 name: "use",
-                aliases: ["use-items"],
+                aliases: ["equip"],
                 memberPerms: [],
                 botPerms: [],
                 requirements: {
@@ -35,15 +35,15 @@ module.exports = class CMD extends Command {
             return message.reply(t("cmds:use.youDontHaveAny"));
         if (!args[1]) args[1] = 1;
         const [item, count] = args;
-        if (userDB.inventory[item] >= count)
+        if (userDB.inventory[item] < count)
             return message.reply(t("cmds:use.tooMuch", { item, count }));
         userDB.inventory[item] = userDB.inventory[item] - count;
         userDB.markModified(`inventory.${item}`);
         let metadata;
         switch (item) {
             case "banknote":
-                userDB.bankLimit = userDB.bankLimit + 100;
-                metadata = t("cmds:use.banknote", { added: 100 });
+                userDB.bankLimit = userDB.bankLimit + 100 * count;
+                metadata = t("cmds:use.banknote", { added: 100 * count });
                 break;
             default:
                 throw new Error(