+++ src/commands/general/listemojis.js
@@ -5,6 +5,7 @@
  */
 const { Permissions } = require("discord.js");
 const { Embed, Command } = require("../../classes");
+const { Pagination } = require("djs-pagination-buttons");
 module.exports = class CMD extends Command {
     constructor(client) {
         super(
@@ -24,126 +25,40 @@ module.exports = class CMD extends Command {
         );
     }
 
-    async execute({ message, args }, t) {
-        if (message.channel.type !== "DM") {
-            const botPerms = message.guild.me.permissionsIn(message.channel);
-            if (!botPerms || !botPerms.has(Permissions.FLAGS.MANAGE_MESSAGES))
-                message
-                    .reply(
-                        `${t("errors:note")}: ${t(
-                            "errors:iDontHavePermission",
-                            {
-                                permission: t("permissions:MANAGE_MESSAGES"),
-                            }
-                        )}, ${t("errors:pagination")}`
-                    )
-                    .then((msg) => {
-                        setTimeout(() => {
-                            msg.delete();
-                        }, 5000);
-                    });
-        }
-        const emojiList = {
-            first: "⏮",
-            back: "⏪",
-            forward: "⏩",
-            last: "⏭",
-            stop: "⏹",
-        };
-        let page = 0;
+    execute({ message }, t) {
+        const pages = [];
         let i0 = 0; //From
         let i1 = 10; //To
-        const embed = new Embed({ color: "green", timestamp: true }).setTitle(
-            t("cmds:listemojis.cmdDesc")
-        );
         const timeout = 200000; //20 secs timeout
         const emojis = message.guild.emojis.cache.size;
         const getList = () => {
             return message.guild.emojis.cache
-                .map((e, x) => `• ${e.name} (${x})`)
+                .map((e, x) => `• ${e.name} (${x}) [\`${e}\`]`)
                 .slice(i0, i1)
                 .join("\n");
         };
-        const curPage = await message.reply({
-            embeds: [
-                embed
-                    .setDescription(`Emojis: ${emojis}\n\n${getList()}`)
-                    .setFooter(
-                        `${t("misc:page")} ${page + 1} / ${Math.ceil(
-                            emojis / 10
-                        )}`
-                    ),
-            ],
-        });
-        for (var key in emojiList) {
-            await curPage.react(emojiList[key]);
-        }
-        const reactionCollector = curPage.createReactionCollector(
-            (reaction, user) =>
-                Object.values(emojiList).includes(reaction.emoji.name) &&
-                user.id === message.author.id,
-            { time: timeout }
-        );
-        reactionCollector.on("collect", (reaction) => {
-            // Remove the reaction when the user react to the message
-            reaction.users.remove(message.author);
-            switch (reaction.emoji.name) {
-                case emojiList["first"]:
-                    i0 = 0;
-                    i1 = 10;
-                    page = 0;
-                    break;
-                case emojiList["last"]:
-                    i0 = Math.floor(emojis / 10);
-                    i1 = Math.floor(emojis / 10) + 10;
-                    page = emojis;
-                    break;
-                case emojiList["back"]:
-                    i0 = i0 - 10;
-                    i1 = i1 - 10;
-                    page = page - 1;
-                    break;
-                case emojiList["forward"]:
-                    i0 = i0 + 10;
-                    i1 = i1 + 10;
-                    page = page + 1;
-                    break;
-                case emojiList["stop"]:
-                    return curPage.delete();
-                    break;
-            }
-            // If there is no emoji to display, delete the message
+        for (let i = 0; i < emojis; ) {
+            const embed = new Embed({ timestamp: true })
+                .setTitle(t("cmds:listemojis.cmdDesc"))
+                .setDesc(getList());
+            const p = pages.length;
+            pages[p] = embed;
+            i = i + 10;
+            i0 = i0 + 10;
+            i1 = i1 + 10;
             if (i1 > emojis + 10) {
-                return curPage.delete();
+                break;
             }
             if (!i0 || !i1) {
-                return curPage.delete();
+                delete pages[p];
+                break;
             }
-            curPage.edit({
-                embeds: [
-                    embed
-                        .setDescription(`Emojis: ${emojis}\n\n${getList()}`)
-                        .setFooter(
-                            `${t("misc:page")} ${page + 1} / ${Math.ceil(
-                                emojis / 10
-                            )}`
-                        ),
-                ],
-            });
+        }
+        const pagination = new Pagination(this.client, {
+            timeout: timeout,
         });
-        reactionCollector.on("end", () => {
-            curPage.reactions.removeAll().catch((err) => {
-                console.error(err);
-            });
-            curPage.edit({
-                embeds: [
-                    embed.setFooter(
-                        `${t("misc:page")} ${page + 1} / ${Math.ceil(
-                            emojis / 10
-                        )} | ${t("misc:ptimeout")}`
-                    ),
-                ],
-            });
-        });
+        pagination.setPages(pages);
+        pagination.setAuthorizedUsers([message.author.id]);
+        pagination.send(message);
     }
 };