+++ src/commands/general/serverinfo.js
@@ -25,9 +25,10 @@ module.exports = class CMD extends Command {
 
     async execute({ message, args }, t) {
         args[1] = args[1] ? args[1].toLowerCase() : "";
+        const { guild } = message;
         const embed = new Embed({ color: "green", timestamp: true })
             .setTitle(t("misc:sinfo"))
-            .setDesc(`${message.guild.id}`)
+            .setDesc(`${guild.description ?? guild.id}`)
             .setThumbnail(message.guild.iconURL());
         const iconURL = message.guild.iconURL().slice(0, 35) + "...";
         message.guild.members.fetch();