+++ src/commands/mod/mute.js
@@ -4,7 +4,6 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const { Permissions } = require("discord.js");
-const { userFromMention } = require("../../helpers/Util.js");
 const { Embed, Command } = require("../../classes");
 module.exports = class CMD extends Command {
     constructor(client) {
@@ -20,13 +19,27 @@ module.exports = class CMD extends Command {
                 disabled: false,
                 cooldown: 10,
                 category: "Moderation",
+                supportsSlash: true,
+                options: [
+                    {
+                        name: "user",
+                        description: "User to mute",
+                        type: "USER",
+                        required: true,
+                    },
+                    {
+                        name: "reason",
+                        description: "Why do you want to mute this guy?",
+                        type: "STRING",
+                        required: false,
+                    },
+                ],
             },
             client
         );
     }
 
     async execute({ message, args, guildDB }, t) {
-        //TODO: Add translation
         const reason = args.slice(1).join(" ") || t("misc:not_spec");
         const user = await this.getUserFromIdOrMention(args[0]);
         if (!user) {
@@ -36,8 +49,7 @@ module.exports = class CMD extends Command {
             return message.reply(t("cmds:mute.errorYourself"));
         }
 
-        let member;
-        member = message.guild.members.cache.get(user.id);
+        let member = message.guild.members.cache.get(user.id);
         if (!member) {
             member = await message.guild.members.fetch(user.id);
             if (!member) return message.reply(t("errors:userNotInGuild"));
@@ -49,32 +61,9 @@ module.exports = class CMD extends Command {
                 })
             );
         }
-        let muteRole = message.guild.roles.cache.find(
-            (r) => r.name === "Muted"
-        );
+        const muteRole = await this.muteRole(message);
         if (message.guild.me.roles.highest.position >= muteRole.position)
-            return message.channel.send(t("misc:higherRoleBot"));
-        if (!muteRole) {
-            try {
-                muteRole = await message.guild.roles.create({
-                    name: "Muted",
-                    color: "#ff0000",
-                    permissions: [],
-                });
-                message.guild.channels.fetch().then((channels) => {
-                    channels.forEach(async (channel) => {
-                        await channel.permissionOverwrites.create(muteRole, {
-                            VIEW_CHANNEL: true,
-                            SEND_MESSAGES: false,
-                            ADD_REACTIONS: false,
-                        });
-                    });
-                });
-            } catch (e) {
-                throw e;
-            }
-        }
-        const embed = new Embed({ color: "error", timestamp: true });
+            return message.reply(t("misc:higherRoleBot"));
         member.roles
             .add(
                 muteRole,
@@ -91,25 +80,110 @@ module.exports = class CMD extends Command {
                     })
                 );
                 if (guildDB.plugins.modlogs) {
-                    const channel = message.guild.channels.cache.get(
-                        guildDB.plugins.modlogs
-                    );
-                    if (channel) {
-                        embed.setTitle(
-                            `${t("cmds:mute.muted")}: ${user.tag} (${user.id})`
-                        );
-                        embed.addField(
-                            t("misc:resMod"),
-                            `${message.author.tag} (${message.author.id})`
-                        );
-                        embed.addField(t("misc:reason"), reason);
-                        channel.send({ embeds: [embed] });
-                    }
+                    this.handleModLogs(message, guildDB, user, reason, t);
                 }
                 message.reply(t("cmds:mute.success", { tag: user.tag }));
             })
             .catch((e) => {
                 throw e;
             });
+    }
+
+    async run({ interaction, guildDB }, t) {
+        const user = interaction.options.getUser("user", true);
+        const reason =
+            interaction.options.getString("reason") ?? t("misc:not_spec");
+        if (!user) {
+            return interaction.followUp(t("errors:invalidUser"));
+        }
+        if (interaction.user.id === user.id) {
+            return interaction.followUp(t("cmds:mute.errorYourself"));
+        }
+
+        let member = interaction.guild.members.cache.get(user.id);
+        if (!member) {
+            member = await interaction.guild.members.fetch(user.id);
+            if (!member)
+                return interaction.followUp(t("errors:userNotInGuild"));
+        }
+        if (member.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES)) {
+            return interaction.followUp(
+                t("cmds:mute.memberHasPerm", {
+                    permission: t("permissions:MANAGE_MESSAGES"),
+                })
+            );
+        }
+        const muteRole = await this.muteRole(interaction);
+        if (interaction.guild.me.roles.highest.position >= muteRole.position)
+            return interaction.followUp(t("misc:higherRoleBot"));
+        member.roles
+            .add(
+                muteRole,
+                t("cmds:mute.reason", {
+                    tag: interaction.user.tag,
+                    reason: reason,
+                })
+            )
+            .then((m) => {
+                m.user.send(
+                    t("cmds:mute.DMtext", {
+                        tag: interaction.user.tag,
+                        reason: reason,
+                    })
+                );
+                if (guildDB.plugins.modlogs) {
+                    this.handleModLogs(interaction, guildDB, user, reason, t);
+                }
+                interaction.followUp(t("cmds:mute.success", { tag: user.tag }));
+            })
+            .catch((e) => {
+                throw e;
+            });
+    }
+
+    async muteRole(message) {
+        let muteRole = message.guild.roles.cache.find(
+            (r) => r.name === "Muted"
+        );
+        if (!muteRole) {
+            try {
+                muteRole = await message.guild.roles.create({
+                    name: "Muted",
+                    color: "#ff0000",
+                    permissions: [],
+                });
+                message.guild.channels.fetch().then((channels) => {
+                    channels.forEach(async (channel) => {
+                        await channel.permissionOverwrites
+                            .create(muteRole, {
+                                VIEW_CHANNEL: true,
+                                SEND_MESSAGES: false,
+                                ADD_REACTIONS: false,
+                            })
+                            .catch(() => {});
+                    });
+                });
+            } catch (e) {
+                throw e;
+            }
+        }
+        return muteRole;
+    }
+
+    handleModLogs(message, guildDB, user, reason, t) {
+        message.author = message.author ?? message.user;
+        const channel = message.guild.channels.cache.get(
+            guildDB.plugins.modlogs
+        );
+        const embed = new Embed({ color: "error", timestamp: true });
+        if (channel) {
+            embed.setTitle(`${t("cmds:mute.muted")}: ${user.tag} (${user.id})`);
+            embed.addField(
+                t("misc:resMod"),
+                `${message.author.tag} (${message.author.id})`
+            );
+            embed.addField(t("misc:reason"), reason);
+            channel.send({ embeds: [embed] });
+        }
     }
 };