+++ src/commands/mod/prune.js
@@ -11,8 +11,8 @@ module.exports = class CMD extends Command {
             {
                 name: "prune",
                 aliases: ["purge"],
-                memberPerms: [Permissions.FLAGS.KICK_MEMBERS],
-                botPerms: [Permissions.FLAGS.KICK_MEMBERS],
+                memberPerms: [Permissions.FLAGS.MANAGE_MESSAGES],
+                botPerms: [Permissions.FLAGS.MANAGE_MESSAGES],
                 requirements: {
                     args: true,
                     guildOnly: true,
@@ -38,7 +38,7 @@ module.exports = class CMD extends Command {
     }
 
     //eslint-disable-next-line no-unused-vars
-    execute({ message, args, guildDB }, t) {
+    async execute({ message, args, guildDB }, t) {
         //TODO: Add translation
         let messages;
         const errMsg =
@@ -48,16 +48,18 @@ module.exports = class CMD extends Command {
                 args[0] = 99;
                 break;
             case "bots":
-                messages = message.channel.messages.cache.filter(
-                    (m) => m.author.bot
-                );
+                messages = await message.channel.messages
+                    .fetch({ limit: 100 })
+                    .then((msgs) => msgs.filter((m) => m.author.bot));
                 break;
         }
         if (typeof args[0] === "string" && args[0].startsWith("*")) {
             args[0] = args[0].slice(1); //Remove * from it
-            messages = message.channel.messages.cache.filter(
-                (m) => m.content.indexOf(args[0]) !== -1
-            );
+            messages = await message.channel.messages
+                .fetch({ limit: 100 })
+                .then((msgs) =>
+                    msgs.filter((m) => m.content.indexOf(args[0]) !== -1)
+                );
         }
         if (!isNaN(parseInt(args[0]))) {
             const amount = parseInt(args[0]) + 1;
@@ -74,13 +76,22 @@ module.exports = class CMD extends Command {
                 );
             }
 
-            message.channel.bulkDelete(amount, true).catch((err) => {
+            messages = await message.channel.messages
+                .fetch({ limit: 100 })
+                .then((msgs) => {
+                    if (args[1] && !args[1] === "-f")
+                        return msgs.filter((msg) => !msg.pinned);
+                    return msgs;
+                });
+            message.channel.bulkDelete(message).catch((err) => {
                 message.client.logger.log(err, "error", ["PRUNING"]);
                 return message.channel.send(errMsg);
             });
         } else if (messages) {
             message.delete();
-            message.channel.bulkDelete(messages, true).catch((err) => {
+            if (args[1] && !args[1] === "-f")
+                messages = messages.filter((msg) => !msg.pinned);
+            message.channel.bulkDelete(messages).catch((err) => {
                 message.client.logger.log(
                     "Error when PRUNING messages",
                     "error",