+++ CHANGELOG.md
@@ -47,7 +47,9 @@ The displayed date is in the format `DD-MM-YYYY`
 - feat: New cmds `chanid`, `roleid`, `serverid`, `userid`
 - feat: New cmd `autorole` & new plugin autorole
 - feat: New cmd `reverse`
-- feat: New slash cmd `ping`
+- feat: New slash commands `ping`, `staff`, `mute`
+- feat: New cmd `config`
+- feat: Add `-f` option for prune cmd
 
 ### Changes